Rework build miner
- Renamed Miner.py to BuildMiner.py.
- Added Miner.py.
- Added builds.txt.

BuildMiner works mostly the same as before, except now it also
associates builds with each SStuB. It is also executed slightly
differently than before.

Miners purpose is to act as a 'hub' for coordinating the mining process,
and has the responsibility of creating and passing shared objects.

Builds.txt stores a list of build files we're looking for.