Fix exception crashing
Various things can go wrong on GitHubs end, resulting in the program
crashing. It seems to simplest way of dealing with this is to just
restart the program whenever an exception occurs, at which point
execution just continues as if nothing ever happened.

Also adjusted the counter so it prints values from the previous
iterations, rather than just the current one.

Fixed an issue in GithubManager where 'exceeded_request_limit' would
never be true if the offset is 0.